[Add]guest_user function

+++ db/migrate/20221128031214_create_posts.rb
@@ -5,8 +5,8 @@ class CreatePosts < ActiveRecord::Migration[6.1]
       t.string :shop_name
       t.text :shop_explanation
       t.string :address
-      t.integer :latitude
-      t.integer :longitude
+      t.float :latitude
+      t.float :longitude
       t.boolean :is_released, null: false, default: "true"
 
       t.timestamps
